Cleanup and reorganize drag/drop/resize handlers
