+++ assets/scripts/app/event_listeners.js
@@ -4,16 +4,12 @@ import {
   onBodyMouseUp
 } from '../segments/drag_and_drop'
 import { onStorageChange } from '../users/authentication'
-import { onGlobalKeyDown } from './keyboard_commands'
 import { onResize } from './window_resize'
 
 export function addEventListeners () {
   window.addEventListener('storage', onStorageChange)
-
   window.addEventListener('resize', onResize)
-
   window.addEventListener('pointerdown', onBodyMouseDown)
   window.addEventListener('pointermove', onBodyMouseMove)
   window.addEventListener('pointerup', onBodyMouseUp)
-  window.addEventListener('keydown', onGlobalKeyDown)
 }
